separate ComposerWindowSize
