+++ extension/chrome/elements/reply_pubkey_mismatch.ts
@@ -108,13 +108,13 @@ Catch.try(async () => {
       Xss.sanitizeAppend('.email_preview', Ui.e('span', { class: 'email_address display_when_sign', text: recipient }));
     }
     $('.pubkey_file_name').text(att.name);
-    composer.resizeComposeBox();
+    composer.composerWindowSize.resizeComposeBox();
     BrowserMsg.send.scrollToElement(parentTabId, { selector: `#${frameId}` });
     $('#input_text').focus();
 
     Catch.setHandledTimeout(() => {
-      $(window).resize(Ui.event.prevent('veryslowspree', () => composer.resizeComposeBox()));
-      $('#input_text').keyup(Ui.event.prevent('slowspree', () => composer.resizeComposeBox()));
+      $(window).resize(Ui.event.prevent('veryslowspree', () => composer.composerWindowSize.resizeComposeBox()));
+      $('#input_text').keyup(Ui.event.prevent('slowspree', () => composer.composerWindowSize.resizeComposeBox()));
     }, 1000);
   };
